W-16354625: Created metadata folder & moved providers accordingly
