+++ src/main/java/org/mule/extension/mulechain/internal/helpers/FileType.java
@@ -1,0 +1,5 @@
+package org.mule.extension.mulechain.internal.helpers;
+
+public enum FileType {
+  PDF, TEXT, URL
+}
