+++ src/main/java/org/mule/extension/mulechain/internal/helpers/FileTypeEmbeddingProvider.java
@@ -4,7 +4,6 @@ package org.mule.extension.mulechain.internal.helpers;
 import java.util.Arrays;
 import java.util.Set;
 
-import org.mule.extension.mulechain.internal.llm.FileType;
 import org.mule.runtime.api.value.Value;
 import org.mule.runtime.extension.api.values.ValueBuilder;
 import org.mule.runtime.extension.api.values.ValueProvider;
