--- src/main/java/org/mule/extension/mulechain/internal/llm/FileType.java
@@ -1,5 +1,0 @@
-package org.mule.extension.mulechain.internal.llm;
-
-public enum FileType {
-  PDF, TEXT, URL
-}
