+++ src/main/java/org/mule/extension/mulechain/internal/llm/type/LangchainLLMType.java
@@ -1,4 +1,4 @@
-package org.mule.extension.mulechain.internal.llm;
+package org.mule.extension.mulechain.internal.llm.type;
 
 public enum LangchainLLMType {
   OPENAI, GROQAI_OPENAI, MISTRAL_AI, OLLAMA, ANTHROPIC, AZURE_OPENAI
