+++ src/main/java/org/mule/extension/mulechain/internal/metadata/TokenStreamMetadataResolver.java
@@ -1,4 +1,4 @@
-package org.mule.extension.mulechain.internal.streaming;
+package org.mule.extension.mulechain.internal.metadata;
 
 import org.mule.metadata.api.model.MetadataType;
 import org.mule.runtime.api.metadata.MetadataContext;
@@ -6,9 +6,9 @@ import org.mule.runtime.api.metadata.resolving.OutputTypeResolver;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class TokenStreamOutputResolver implements OutputTypeResolver<String> {
+public class TokenStreamMetadataResolver implements OutputTypeResolver<String> {
 
-  private static final Logger LOGGER = LoggerFactory.getLogger(TokenStreamOutputResolver.class);
+  private static final Logger LOGGER = LoggerFactory.getLogger(TokenStreamMetadataResolver.class);
 
   @Override
   public MetadataType getOutputType(MetadataContext metadataContext, String key) {
